app.py provides the basic streamlit application and end user controls.
locationutils.py provides the Geocoder class, an interface for the nominatim API. The PhysicalLocation class receives the output from the Geocoder.reverse() method (json dict) and normalizes each location with None values when the API doesn't provide data. This is all done in the constructor for PhysicalLocation.

Once the resolved location is normalized, it's added to the pandas dataframe in app.py. When all locations are resolved, the dataframe is exported to a .csv for download.